Move uncertain objects into top sdk package
