+++ sdk/src/main/scala/sdk/Codec.scala
@@ -1,4 +1,6 @@
-package sdk.data
+package sdk
+
+import sdk.data.ByteArray
 
 trait Codec[F[_], A] {
   def encode(x: A): F[ByteArray]
