+++ sdk/src/main/scala/sdk/data/Blake2b256Hash.scala
@@ -1,6 +1,7 @@
 package sdk.data
 
 import blakehash.Blake2b256
+import sdk.Base16
 
 import java.io.OutputStream
