+++ sdk/src/main/scala/sdk/data/CodecBlake2b256Hash.scala
@@ -2,6 +2,7 @@ package sdk.data
 
 import cats.Applicative
 import cats.syntax.all.*
+import sdk.Codec
 
 class CodecBlake2b256Hash[F[_]: Applicative] extends Codec[F, Blake2b256Hash] {
   def encode(x: Blake2b256Hash): F[ByteArray] = x.bytes.pure
