+++ sdk/src/main/scala/sdk/data/CodecByteArray.scala
@@ -2,6 +2,7 @@ package sdk.data
 
 import cats.Applicative
 import cats.syntax.all.*
+import sdk.Codec
 
 class CodecByteArray[F[_]: Applicative] extends Codec[F, ByteArray] {
   def encode(x: ByteArray): F[ByteArray] = x.pure
